Customizing build settings by notebook kernel
This is essential because the R kernel produces LaTeX output that MathJax can't handle (e.g., tabular environments that aren't part of math mode).  So we have to disable LaTeX for R.  (But not for Python so that SymPy can still do nice math output.)  This actually makes outputs that look more like RStudio notebooks anyway.